Refactors to abstract additional code.
